src/main/java/Shop/GuitarStrings.java
package Shop;

public class GuitarStrings implements ISell {

    private String description;
    private double buyingPrice;
    private double sellingPrice;

    public GuitarStrings(String description, double buyingPrice, double sellingPrice) {
        this.description = description;
        this.buyingPrice = buyingPrice;
        this.sellingPrice = sellingPrice;
    }

    public String getDescription() {
        return description;
    }

    public double getBuyingPrice() {
        return buyingPrice;
    }

    public double getSellingPrice() {
        return sellingPrice;
    }

    public double calculateMarkUp() {
        return this.sellingPrice - this.buyingPrice;
    }
}
